renamed utility func
